Account for the whole day
In order to know whether it's my birthday we really need to check that
it's between a timestamp at the start of the day and a timestamp at
the end of the day.

If `t` is midnight of a particular day, is `t + 1.day` always the next
day? An interesting question.

For now, though, we have to adjust for the fact that I do not live in
UTC.