Use time zones
I live in chicago so we need to interpret WHEN_IS_MY_BIRTHDAY in
Chicago time rather than in UTC. So the timestamp I'm working with
becomes one without a specified timezone and I use a particular one to
turn it into a moment in time that I can use to generate a range.

But what if I'm somewhere else for my birthday?