Support my local timezone
Because my birthday is where I am (I have an ego, so what :P) and I
can move around, I need to be able to have this code check whether the
current time falls within the range that is my birthday in my local
timezone. What this means is that now I store WHEN_IS_MY_BIRTHDAY as a
string and then parse it with the passed tz. Then use the instant in
time denoted by parsing that time string in the passed tz to construct
an interval and check that the current local time is within it.

This code is for a pretty trivial example, but it's already getting
kinda hairy. I need to remember whenever I interpret
WHEN_IS_MY_BIRTHDAY to use the timezone associated with the person
who's birthday it is, and I have to be careful about all the usual
errors with things falling inside of ranges.

Is there a better way?