Use a Date instead of a Time
My birthday is not a time, nor is it a time range. My birthday is a
date. So the solution here is simple. Ask me what day it is where I
am and then compare it to the date that I know is my birthday. This
code is almost as simple as the most trivial possible code that we
wrote using Time, but it handles _all_ of the edge cases that we had
to add so much code for.

Dates are not suitable for all things. Sometimes you really do need to
denote a range in time (such as the opening hours of a store). However
if what you want to model is something happening on a certain day, you
will incur much pain by trying to use a range in time to represent the
day. Just use a Date.